Export type equality operator
Previous to `ghc-9.6` the type equality operator (`~`) was a compiler
builtin where as in `ghc-9.6` (and presumably later) is is a `type`
operator defined in `Data.Type.Equality` and re-exported by `Prelude`.

The above means that with `ghc-9.6` using a custom prelude (ie
`cardano-prelude`) will mean that `~` is undefined. The soltuion
is to export it.