Reinstate toS function and protolude version of show
The `toS` function is parameterized on both the input and output
types which is kind of shitty from a code maintence POV.

Similarly protoludes `show` function is parametric on both input
and output types. This function is now deprecated and of one
`Prelude.show`, `textShow` or `bsShow` (later 2 newly added) should
be used instead.